adapter: provision new `sequence_explain` paths
Provision two new `Explainee` types:

- `MaterializedView(GlobalId)` for `EXPLAIN MATERIALIZED VIEW`
- `Index(GlobalId)` for `EXPLAIN VIEW`

Also, add a comment that marks `Dataflow(GlobalId)` as "deprecated"
(this variant will be removed when we resolve #18089).

Similarly, provision execution paths in `sequence_explain` to handle
plans with the above `Explainee` variants as follows:

1. Dispatch `MaterializedView(.)` to `explain_materialized_view`.
2. Dispatch `Index(_)` to `explain_index`.
3. Dispatch `Query` and `Dataflow(_)` to `explain_query`.

The methods handling (1) and (2) are new stubs.
The method handling (3) is just `sequence_explain_plan` renamed.