sql-parser: add `impl_simple_options` macro
We have several paths in our SQL parser where we want to generate
options list that  are (a) meant for internal use, (b) subject to
frequent change, and (c) have option names that are all simple
(unquoted) idents.

Define an `impl_simple_options` macro that simplifies the definition of
AST structs for such option lists.